requirements.txt
# Core dependencies
playwright>=1.40.0
pyyaml>=6.0.1
requests>=2.31.0
asyncio-mqtt>=0.13.0

# MCP (Model Context Protocol) dependencies
mcp>=1.0.0
websockets>=11.0.3

# Logging and reporting
structlog>=23.2.0
colorlog>=6.7.0

# Utilities
click>=8.1.7
python-dateutil>=2.8.2
pytz>=2023.3

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-playwright>=0.4.3

# Development tools (optional)
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1

# Performance monitoring (optional)
psutil>=5.9.6
memory-profiler>=0.61.0